Add preference for showing the copy mode
If the variable `@prefix_highlight_show_copy_mode` is set to `"on"`, the
`#{prefix_highlight}` placeholder will be replaced with the string
`Copy` whenever copy mode is enabled. The prefix highlight takes
precedence over showing copy mode. The attributes for copy mode can be
controlled by the `@prefix_highlight_copy_mode_attr` variable, which
should be a comma-separated list of attributes, with the default value
of `fg=default,bg=yellow`.

At the end of the token, use `#[default]` instead of
`#[fg=default,bg=default]`. This is done because the copy mode may
include arbitrary attributes, and also to make it easier in the future
to change the prefix highlight config to use arbitrary attributes too.